art_group: fix handling of folded headers.
When parsing a header only for known fields,
and encountering an unknown field,
the current unfolding buffer was left unchanged
instead of being reset.
It could lead to parts of unknown fields to be
appended to known fields.

Furthermore, the unfolding code would assume that
the unfolding buffer is set,
which is not true if this is the first field and
it is being ignored.